requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
ntplib==0.4.0
schedule==1.2.0
pytz==2023.3
psutil==5.9.6
